+++ src/default.jl
@@ -27,7 +27,7 @@ function default(algo::Function)
     ##
 
     in((nipals, nipalsmiss)).(algo) ? dump(Jchemo.ParNipals()) : nothing
-    in((snipals_soft2, snipals_h, snipals_soft)).(algo) ? dump(Jchemo.ParSnipals()) : nothing
+    in((snipals_softs, snipals_h, snipals_soft)).(algo) ? dump(Jchemo.ParSnipals()) : nothing
 
     in((pcasvd, pcaeigen, pcaeigenk, pcasph)).(algo) ? dump(Jchemo.ParPca()) : nothing
     in((pcanipals, pcanipalsmiss)).(algo) ? dump(Jchemo.ParPcanipals()) : nothing